Add util folder
